# src/main/resources/static/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "tags": [
    {
      "name": "Billing API",
      "description": "API pour la gestion de la facturation des clients de Tech Solutions"
    }
  ],
  "paths": {
    "/invoices/{id}/pay": {
      "put": {
        "tags": [
          "Billing API"
        ],
        "operationId": "payInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/invoices": {
      "post": {
        "tags": [
          "Billing API"
        ],
        "operationId": "createInvoice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/{id}": {
      "get": {
        "tags": [
          "Billing API"
        ],
        "operationId": "getInvoice",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/invoices/pending": {
      "get": {
        "tags": [
          "Billing API"
        ],
        "operationId": "getPendingInvoices",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 15
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}/total": {
      "get": {
        "tags": [
          "Billing API"
        ],
        "operationId": "getClientTotal",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TotalAmountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}/invoices": {
      "get": {
        "tags": [
          "Billing API"
        ],
        "operationId": "getClientInvoices",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvoiceResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "InvoiceStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          }
        }
      },
      "CreateInvoiceRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "InvoiceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "clientId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number"
          },
          "description": {
            "type": "string"
          },
          "dateEmission": {
            "type": "string",
            "format": "date"
          },
          "datePaiement": {
            "type": "string",
            "format": "date"
          },
          "status": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          }
        }
      },
      "TotalAmountResponse": {
        "type": "object",
        "properties": {
          "totalAmount": {
            "type": "number"
          }
        }
      }
    }
  }
}